add option to run local devserver
this is needed because a 'stupid' server would
now look for files that have the name of the city in the url.

So when we go to `/berlin` it would search for `berlin.html`, not find
it and show a 404. This server allows us to always serve `index.html` 
if the file is not found and let javascript handle the city resolution.